requirements.txt
flask==1.1.2
pip==19.2.3
numpy==1.21.5
pandas==1.3.5
Flask-SQLAlchemy==2.5.1
flask_migrate==3.1.0
pillow==8.4.0
cmapy==0.6.6
